feat(utils): export PartialDeep and ReadonlyDeep from type-fest
Type-fest provides two generic types very useful for configuration
object: PartialDeep and ReadonlyDeep. Since they are quite complex,
it is better to rely on a proven library instead of trying to recreate
the types.
Also, instead of installing type-fest in multiple places, I think it is
easier to handle all the generic types from the utils package.