feat(ui): add a helper for theme creation to build the shadows
The ShadowTokens type is quite nested and it is easy to reach a file
with more than 500 lines to create the shadows. So it becomes less
readable and it offers an unpleasant experience for the theme designer.
The goal of the `getShadowTokens` helper is to build the theme shadows
by providing only a few values.