refactor(ui): remove Field component
Because Field component was accepting either an input type or a
textarea, it was causing Typescript errors on Storybook stories. It is
not really useful. The consumer should import directly the desired
component.
Also, it could be confusing when using the LabelledField component
since the `field` attribute is not restricted to the Field component.
I also moved all the field components in a separate directory. This way
we can reuse the style for inputs, textarea and select.

BREAKING CHANGE: Field, FieldType and FieldProps have been removed