fix(ui): make List spacing nullable
By default, the list item was spaced when using a regular list and not
spaced when using a bordered list. For design purposes, it can be
useful to remove the spacing or to add a spacing to bordered lists. So
instead the spacing attribute is now nullable with a default. The
consumer should decide if the spacing should be removed or not.

BREAKING CHANGE: List component has now a nullable default spacing.