fix(ui): allow consumer to style Collapsible parts
Depending on context, the collapsible body should not be bordered
by default. Also, the consumer might need to add custom styles
such as adding padding to the body and/or summary.
In addition, the Collapsible component should not be responsible
of the icon. In some context, it could be acceptable to use a
plus/minus icon instead of an angle.

BREAKING CHANGE: You need to provide an icon to the Collapsible
component.