fix(ui): use React ref as useScrollbarWidth parameter
We should not use a querySelector to pass a HTMLElement to the hook and
if we pass directly `ref.current` it does not always update the
scrollbar width. So the expected parameter is now a RefObject.
I also updated the getScrollbarWidth helper to receive an HTMLElement
instead of a querySelector string.
I also added documentation and tests for the hook.

BREAKING CHANGE: expected parameter in useScrollbarWidth and
getScrollbarWidth has changed.