feat(ui): use themes contract to set the Drawer max width
Instead of using an attribute to define the max width used by the
Drawer component, we use the new token `size.sidebar` from themes
contract. It allows themes designer to control the size of the drawer
instead of having a fixed value in source code.

BREAKING CHANGE: the `maxWidth` attribute of Drawer component has been
removed.