fix(ui): remove empty CSS vars from style prop
In SSR mode, when assigning an empty string to a CSS var, the var is
removed on server-side but restore on client-side. Because of that,
React will print an error "Prop `style` did not match".
So instead of conditionally assigning an empty string, we
conditionnally add the CSS var to the `assignInlineVars` function.