feat(ui): export dark/light themes separately
It can be useful for the consumer to retrieve the theme ids with
correct type inference. By filtering by scheme, Typescript infers the
wrong type even with a type guard.
So instead we use three different arrays: one for the dark themes, a
second for the light themes and another ones for all themes.
Now we have three types based on the id of the existing themes. They
can be used to check the correct value in `data-theme` attribute.
I also added three helpers functions to check if a given id matches
an existing theme.

BREAKING CHANGE: the `themes` export is no longer an object but an
array of themes and the `Themes` type no longer exists.