fix(server): use API endpoint to serve API
There was two problems.
First, the use of vite middleware before the loadAPI middleware caused
a 500 error because of headers. I'm not sure why... but switching the
two statements fix the issue.
Then the consumer was forced to declared the API twice: once when
creating the API and once in the server config. It was prone to error.
Since we receive the API instance, we already have the route to use so
we can reuse it.

BREAKING CHANGE: the API config is no longer an object, you must pass
the instance directly and declare the route (if needed) in the API
config.