feat(server): add Request object as parameter in render function
Depending on the consumer method to render the contents, the URL might
not be enough. For example, with `react-router-dom` and its data API,
we need the original request to be able to use SSR.
We could remove the url argument since the request already has this
information. However, we're using it in a Vite method so might as well
leave this parameter.

BREAKING CHANGE: ServerRender type has been renamed to RenderFunction