feat(core): convert router to a data router
Since v6.4, React Router encourages us to convert our app to use a data
router. Some convenients API, like `<Redirect />` component have been
removed. So we was using `<Navigate />` instead but it causes issues
when using SSR.
To fix those issues and to be up to date with the last React Router API
we convert the previous `BrowserRouter` router to a data router using
`createBrowserRouter`.
Also, we no longer need the previous `App` component so, we replace it
with the `Layout` component. Since it is a special component, it should
be apart.