feat(core): add a pages configuration key
Currently we're using the filename as page name and to generate the
slug.

With an English website we could define some reserved filename
for these pages. However if we want to take into account other
languages, it makes no sense for the consumer to have a page called
"Home" if in its language the right word is "Accueil" for example.

We cannot plan a reserved filename for every language. So instead we
allow users to configure the page name for some specific pages like
the homepage or the legal notice page.

The homepage will always appear first in the main navigation and
the legal notice page will only appear in the footer.