feat(ui): replace UIContext with LinkContext
Currently, the UIContext was only used to provide a Link component. In
the future, if we need to override other components, it does not make
sense to use a single context. React looks only for the closest context
provider above. So, if a single component needs to be replaced, we will
have to provide all the components.
The link provider is now accepting a nullable value. This way we can
override the previous link explicitly.

BREAKING CHANGE: The UIContext/UIProvider couple has been replaced with
LinkContext/LinkProvider