refactor(api): rewrite custom exceptions
Two custom exceptions are enough to handle the different kind of
errors:
* the internal ones which should be masked by Yoga
* the input errors which should be communicated to the user
Also we should not test the errors messages but instead the error kind
or the error code. So the messages should not be exported from an
external file but defined in the error. It improve readability and it
is easier to search where the message was defined.