fix(api): rename contents field on DocDirectory to entries
When querying a doc entry, the contents field had to be aliased because
of the different type in DocFile (string) and DocDirectory
(connection). Typescript can't know that we are using an alias so it
was not type-safe. Using two different fields solves this issue.

BREAKING CHANGE: contents field no longer exists in DocDirectory type