feat(api): remove nested pagination on documentation entries
This commit reverts #d6b4d36 which introduced nested pagination. It
could work on single entry/directory (it does not currently) but it
make no sense on connections (entries/directories).
So it is preferable to revert these changes. If the consumer needs
paginated entries, it should do two different requests by omitting the
entries node on DocDirectory. This way it is possible to query all the
entries at once without pagination.

BREAKING CHANGE: nested paginations are no longer available