test(api): fix compatibility issue in fixtures removal
It seems I was using Node v20 when writing the previous version of
deleteFixturesIn so everything was working fine. However, the project
should be compatible with lower version of Node. With Node 18 for
example, `path` does not exist in Dirent object. So I updated the
`deleteFixturesIn` function to support lower Node versions.