feat(server): make SSR rendering agnostic
We were relying on placeholders and HTML file to generate the file
contents. It should be the role of the consumer instead of the server.
It makes it easier for the consumer to update its rendering method (for
example switching from React `renderToString` method to
`renderToPipeableStream` method) and to add more placeholders if
needed.

BREAKING CHANGE: the signature of the render function has been modified
and you no longer need to provide placeholders and the path of an html
file to use SSR. It is up to you to determine how to generate your HTML