feat(core): replace renderToString with renderToPipeableStream method
This is the recommended way with React 18. It adds support for
streaming or waiting for data and we can now use Suspense.
However we should look for a solution to hydrate the entire
document, not only the body. It will ease the hydration of meta
tags. For now, I'm not able to inject the CSS generated by vanilla
extract when hydrating the html so this is the only working solution
I found.