fix(utils): avoid conflict in browser because of node built-in modules
The paths helpers import `fs/promises`. These functions should only be
executed in Node context. However, when we bundle the import became the
first statement and it causes errors in browsers since Node built-in
modules are not available.
To avoid these issues, we should use dynamic import. The `fs/promises`
module will be loaded only if one of the paths helpers is called.