fix(ui): let consumers take care of building themes
We were shipping all the themes so once the package imported in the
consumer app, he only had to add a `data-theme` attribute. For now,
we only have two themes so it is not an issue. However if the themes
library increases, the consumer will have a lot of unused CSS.
Instead we should let the consumer build the required themes. This way
if only one theme is needed, they only get the CSS variables associated
to this theme.